feat: Encrypt oauth data in the store
